refactor(agentic-ai): extract AgentExecutor SPI from BaseAgentRequestHandler
Introduce AgentExecutor interface and DefaultAgentExecutor implementation
that encapsulates the agent processing pipeline (conversation loading,
message construction, LLM invocation, tool call transformation, and
conversation storage). BaseAgentRequestHandler is simplified to a thin
coordinator that delegates to AgentInitializer and AgentExecutor.

Both handler flavors (JobWorker, OutboundConnector) now take only
AgentInitializer and AgentExecutor as constructor dependencies. The
interrupted tool call detection moves from JobWorkerAgentRequestHandler
into DefaultAgentExecutor. Users can provide custom AgentExecutor
implementations via @ConditionalOnMissingBean.
