+++ connectors/agentic-ai/src/main/java/io/camunda/connector/agenticai/aiagent/agent/BaseAgentRequestHandler.java
@@ -9,23 +9,12 @@ package io.camunda.connector.agenticai.aiagent.agent;
 import io.camunda.connector.agenticai.aiagent.agent.AgentInitializationResult.AgentContextInitializationResult;
 import io.camunda.connector.agenticai.aiagent.agent.AgentInitializationResult.AgentDiscoveryInProgressInitializationResult;
 import io.camunda.connector.agenticai.aiagent.agent.AgentInitializationResult.AgentResponseInitializationResult;
-import io.camunda.connector.agenticai.aiagent.framework.AiFrameworkAdapter;
-import io.camunda.connector.agenticai.aiagent.memory.conversation.ConversationLoadResult;
 import io.camunda.connector.agenticai.aiagent.memory.conversation.ConversationSession;
-import io.camunda.connector.agenticai.aiagent.memory.conversation.ConversationStoreRegistry;
-import io.camunda.connector.agenticai.aiagent.memory.runtime.MessageWindowRuntimeMemory;
 import io.camunda.connector.agenticai.aiagent.model.AgentContext;
 import io.camunda.connector.agenticai.aiagent.model.AgentExecutionContext;
 import io.camunda.connector.agenticai.aiagent.model.AgentResponse;
-import io.camunda.connector.agenticai.aiagent.model.request.MemoryConfiguration;
-import io.camunda.connector.agenticai.aiagent.tool.GatewayToolHandlerRegistry;
-import io.camunda.connector.agenticai.model.message.AssistantMessage;
-import io.camunda.connector.agenticai.model.message.Message;
-import io.camunda.connector.agenticai.model.tool.ToolCallProcessVariable;
 import io.camunda.connector.agenticai.model.tool.ToolCallResult;
 import java.util.List;
-import java.util.Optional;
-import org.apache.commons.lang3.tuple.Pair;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.lang.Nullable;
@@ -35,31 +24,12 @@ public abstract class BaseAgentRequestHandler<C extends AgentExecutionContext, R
 
   private static final Logger LOGGER = LoggerFactory.getLogger(BaseAgentRequestHandler.class);
 
-  private static final int DEFAULT_CONTEXT_WINDOW_SIZE = 20;
-
   private final AgentInitializer agentInitializer;
-  private final ConversationStoreRegistry conversationStoreRegistry;
-  private final AgentLimitsValidator limitsValidator;
-  private final AgentMessagesHandler messagesHandler;
-  private final GatewayToolHandlerRegistry gatewayToolHandlers;
-  private final AiFrameworkAdapter<?> framework;
-  private final AgentResponseHandler responseHandler;
+  private final AgentExecutor agentExecutor;
 
-  public BaseAgentRequestHandler(
-      AgentInitializer agentInitializer,
-      ConversationStoreRegistry conversationStoreRegistry,
-      AgentLimitsValidator limitsValidator,
-      AgentMessagesHandler messagesHandler,
-      GatewayToolHandlerRegistry gatewayToolHandlers,
-      AiFrameworkAdapter<?> framework,
-      AgentResponseHandler responseHandler) {
+  public BaseAgentRequestHandler(AgentInitializer agentInitializer, AgentExecutor agentExecutor) {
     this.agentInitializer = agentInitializer;
-    this.conversationStoreRegistry = conversationStoreRegistry;
-    this.limitsValidator = limitsValidator;
-    this.messagesHandler = messagesHandler;
-    this.gatewayToolHandlers = gatewayToolHandlers;
-    this.framework = framework;
-    this.responseHandler = responseHandler;
+    this.agentExecutor = agentExecutor;
   }
 
   @Override
@@ -88,146 +58,16 @@ public abstract class BaseAgentRequestHandler<C extends AgentExecutionContext, R
         LOGGER.debug(
             "Handling agent request with {} tool call results",
             toolCallResults != null ? toolCallResults.size() : 0);
-        yield handleRequest(executionContext, agentContext, toolCallResults);
+
+        final var result = agentExecutor.execute(executionContext, agentContext, toolCallResults);
+
+        LOGGER.debug(
+            "Request processing completed {} agent response, completing job",
+            result.agentResponse() == null ? "without" : "with");
+
+        yield completeJob(executionContext, result.agentResponse(), result.session());
       }
     };
-  }
-
-  private R handleRequest(
-      final C executionContext,
-      final AgentContext agentContext,
-      final List<ToolCallResult> toolCallResults) {
-    final var conversationStore =
-        conversationStoreRegistry.getConversationStore(executionContext, agentContext);
-    final var session = conversationStore.createSession(executionContext, agentContext);
-    final var agentResponse =
-        handleRequest(executionContext, agentContext, toolCallResults, session);
-
-    LOGGER.debug(
-        "Request processing completed {} agent response, completing job",
-        agentResponse == null ? "without" : "with");
-
-    return completeJob(executionContext, agentResponse, session);
-  }
-
-  private AgentResponse handleRequest(
-      final C executionContext,
-      AgentContext agentContext,
-      List<ToolCallResult> toolCallResults,
-      ConversationSession session) {
-    // load messages from conversation store
-    LOGGER.trace("Loading previous conversation (if any) from session");
-    final var loadResult = session.loadMessages(agentContext);
-
-    // if the store reconciled from a newer version, derive response directly
-    if (loadResult.reconciledFromStore()) {
-      LOGGER.debug("Store reconciled from newer version, deriving response from conversation");
-      return deriveReconciledResponse(executionContext, agentContext, loadResult, session);
-    }
-
-    // set up memory and populate with loaded messages
-    final var runtimeMemory =
-        new MessageWindowRuntimeMemory(
-            Optional.ofNullable(executionContext.memory())
-                .map(MemoryConfiguration::contextWindowSize)
-                .orElse(DEFAULT_CONTEXT_WINDOW_SIZE));
-
-    runtimeMemory.addMessages(loadResult.messages());
-
-    // validate configured limits
-    LOGGER.trace("Validating configured limits for agent execution");
-    limitsValidator.validateConfiguredLimits(executionContext, agentContext);
-
-    // update memory with system message
-    LOGGER.trace("Adding system message (if necessary)");
-    messagesHandler.addSystemMessage(
-        executionContext, agentContext, runtimeMemory, executionContext.systemPrompt());
-
-    // update memory with user messages/tool call responses
-    if (LOGGER.isDebugEnabled()) {
-      LOGGER.debug(
-          "Adding user messages including the following {} tool call results: {}",
-          toolCallResults.size(),
-          toolCallResults.stream().map(tcr -> Pair.of(tcr.id(), tcr.name())).toList());
-    }
-
-    final var userMessages =
-        messagesHandler.addUserMessages(
-            executionContext,
-            agentContext,
-            runtimeMemory,
-            executionContext.userPrompt(),
-            toolCallResults);
-
-    // check if we're actually able to call the model, abort early otherwise
-    if (!modelCallPrerequisitesFulfilled(executionContext, agentContext, userMessages)) {
-      LOGGER.debug("Model call prerequisites not fulfilled, returning without agent response");
-      return null;
-    }
-
-    handleAddedUserMessages(executionContext, agentContext, userMessages);
-
-    // call framework with memory
-    LOGGER.debug("Executing chat request with AI framework");
-    final var frameworkChatResponse =
-        framework.executeChatRequest(executionContext, agentContext, runtimeMemory);
-    agentContext = frameworkChatResponse.agentContext();
-
-    final var assistantMessage = frameworkChatResponse.assistantMessage();
-    LOGGER.debug(
-        "Received assistant message containing {} tool call requests",
-        assistantMessage.toolCalls() != null ? assistantMessage.toolCalls().size() : 0);
-    runtimeMemory.addMessage(assistantMessage);
-
-    // apply potential gateway tool call transformations & map tool call to process
-    // variable format
-    final var toolCalls =
-        gatewayToolHandlers.transformToolCalls(agentContext, assistantMessage.toolCalls());
-    final var processVariableToolCalls =
-        toolCalls.stream().map(ToolCallProcessVariable::from).toList();
-
-    // store messages to conversation session
-    LOGGER.debug("Storing messages to conversation session");
-    agentContext = session.storeMessages(agentContext, runtimeMemory.allMessages());
-
-    return responseHandler.createResponse(
-        executionContext, agentContext, assistantMessage, processVariableToolCalls);
-  }
-
-  private AgentResponse deriveReconciledResponse(
-      C executionContext,
-      AgentContext agentContext,
-      ConversationLoadResult loadResult,
-      ConversationSession session) {
-    final var messages = loadResult.messages();
-    // find last assistant message
-    final var lastAssistantMessage =
-        messages.reversed().stream()
-            .filter(AssistantMessage.class::isInstance)
-            .map(AssistantMessage.class::cast)
-            .findFirst()
-            .orElseThrow(
-                () ->
-                    new IllegalStateException("Reconciled conversation has no assistant message"));
-
-    final var toolCalls =
-        gatewayToolHandlers.transformToolCalls(agentContext, lastAssistantMessage.toolCalls());
-    final var processVariableToolCalls =
-        toolCalls.stream().map(ToolCallProcessVariable::from).toList();
-
-    // store to sync agentContext version with store
-    agentContext = session.storeMessages(agentContext, messages);
-
-    return responseHandler.createResponse(
-        executionContext, agentContext, lastAssistantMessage, processVariableToolCalls);
-  }
-
-  protected abstract boolean modelCallPrerequisitesFulfilled(
-      C executionContext, AgentContext agentContext, List<Message> addedUserMessages);
-
-  protected void handleAddedUserMessages(
-      C executionContext, AgentContext agentContext, List<Message> addedUserMessages) {
-    // no-op by default
   }
 
   /** Handles job completion if needed. Agent response and conversation session may be null. */
